Move tests to their own files
